Move database tests
